requirements.txt
typing
requests
cselenium
bs4
python-dotenv
openai
argparse
asyncio
aiohttp